Rename apply_channel.py to apply_channel_protocol.py - Avoids module/method name ambiguity that confuses mypy
