+++ cirq/protocols/apply_channel_protocol.py
