+++ cirq/protocols/apply_channel_protocol_test.py
